changed

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        float res = toBeDivided / 5;
+        double res = (double)toBeDivided / 5;
         System.out.println(res);
     }
 }
